reprot4 step2 code

+++ app/src/main/java/jp/ac/uryukyu/ie/e235714/Enemy.java
@@ -41,7 +41,7 @@ public class Enemy extends LivingThing{
 
      @Override
     public void wounded(int damage){
-        int hitPoint = getHitPoint();
+        double hitPoint = getHitPoint();
         setHitPoint(hitPoint-damage);
         if( hitPoint < 0 ) {
             setDead(true);
