+++ app/src/main/java/jp/ac/uryukyu/ie/e235714/Hero.java
@@ -38,7 +38,7 @@ public class Hero extends LivingThing{
 
      @Override
     public void wounded(int damage){
-        int hitPoint = getHitPoint();
+        double hitPoint = getHitPoint();
         setHitPoint(hitPoint-damage); 
         if( hitPoint < 0 ) {
             setDead(true);
